Added more basic functionality, etc.
Please note that this commit might break a couple of things at first:

- You will need a new run configuration as the new entry point is now in OctoBUS in the main source directory. I'm guessing this will also affect Travis. @LHulk might want to look into that.
- You will have to set your language level for both the module and the project to 1.8 (lambdas as well as switch statements for strings are currently being used)
- If there's anything more and you can't figure out what's wrong exaclty: Let me know :)!

As far as functionality is concerned, we now have basic support for a couple of listeners (keep an eye on the console when clicking the big buttons in the MainPanel). Clicking on "Mitarbeiter-Login" will now open up a basic version of the employees' interface (no authentication yet).